Release/1.1.1

+++ agent-ndk/src/main/cpp/signal-handler.cpp
@@ -5,6 +5,7 @@
 
 #include <cstring>
 #include <pthread.h>
+#include <errno.h>
 
 #include <agent-ndk.h>
 #include "signal-utils.h"
@@ -17,6 +18,7 @@ typedef struct observed_signal {
     const char *name;
     const char *description;
     struct sigaction sa_previous;
+    volatile sig_atomic_t intercepting;
 
 } observed_signal_t;
 
@@ -25,10 +27,9 @@ void invoke_sigaction(int signo, struct sigaction *_sigaction, siginfo_t *_sigin
 
 void invoke_previous_sigaction(int signo, siginfo_t *_siginfo, void *context);
 
-/* our handler interceptors */
-void install_handler();
-
 void uninstall_handler();
+
+observed_signal_t *observed_signal_get_or_null(int);
 
 /* replacement stack */
 static stack_t _stack = {};
@@ -38,12 +39,12 @@ static pthread_mutex_t mutex = PTHREAD_MUTEX_INITIALIZER;
 
 /* Collection of observed signals */
 static observed_signal_t observedSignals[] = {
-        {SIGILL,  "SIGILL",  "Illegal instruction",                               {}},
-        {SIGTRAP, "SIGTRAP", "Trap (invalid memory reference)", {}},
-        {SIGABRT, "SIGABRT", "Abnormal termination",                              {}},
-        {SIGFPE,  "SIGFPE",  "Floating-point exception",                          {}},
-        {SIGBUS,  "SIGBUS",  "Bus error (bad memory access)",                     {}},
-        {SIGSEGV, "SIGSEGV", "Segmentation violation (invalid memory reference)", {}},
+        {SIGILL,  "SIGILL",  "Illegal instruction",                               {}, 0},
+        {SIGTRAP, "SIGTRAP", "Trap (invalid memory reference)",                   {}, 0},
+        {SIGABRT, "SIGABRT", "Abnormal termination",                              {}, 0},
+        {SIGFPE,  "SIGFPE",  "Floating-point exception",                          {}, 0},
+        {SIGBUS,  "SIGBUS",  "Bus error (bad memory access)",                     {}, 0},
+        {SIGSEGV, "SIGSEGV", "Segmentation violation (invalid memory reference)", {}, 0},
 };
 
 static size_t observedSignalCnt = sizeof(observedSignals) / sizeof(observedSignals[0]);
@@ -58,52 +59,63 @@ void interceptor(int signo, siginfo_t *_siginfo, void *ucontext) {
     const auto *_ucontext = static_cast<const ucontext_t *>(ucontext);
 
     if (!intercepting++) {
-        char *buffer = new char[BACKTRACE_SZ_MAX];
+        observed_signal_t *signal = observed_signal_get_or_null(signo);
 
-        for (size_t i = 0; i < observedSignalCnt; i++) {
-            if (observedSignals[i].signo == signo) {
-                const observed_signal_t *signal = &observedSignals[i];
-                _LOGD("Signal %d intercepted: %s", signal->signo, signal->description);
-            }
+        if (nullptr == signal) {
+            _LOGE("Can't reference observed_signal element for signal[%d]", signo);
+            return;
         }
 
+        _LOGD("Signal %d intercepted: %s", signal->signo, signal->description);
+
+        signal->intercepting++;
+        _LOGD("Observer for signal[%d] is intercepting [%d callers]", signal->signo,
+              signal->intercepting);
+
+        char *buffer = new char[BACKTRACE_SZ_MAX];
         if (collect_backtrace(buffer, BACKTRACE_SZ_MAX, _siginfo, _ucontext)) {
             serializer::from_crash(buffer, std::strlen(buffer));
         }
-
-        delete [] buffer;
+        delete[] buffer;
 
         // Uninstall the custom handler prior to calling the previous sigaction (to prevent recursion)
         uninstall_handler();
 
         // chain to previous signo handler for abend
         invoke_previous_sigaction(signo, _siginfo, ucontext);
-    }
 
-    intercepting--;
+        signal->intercepting--;
+
+        intercepting--;
+    }
 }
 
 /**
- * Install observed signals
- */
-void *install_handler(__unused void *unused) {
+ * Install observed signal handlers
+ *
+ * SA_ONSTACK: handle the signal on separate stack
+ **/
+void *install_signal_observers(__unused void *unused) {
 
     if (!initialized++) {
-        pthread_setname_np(pthread_self(), "NR-Signal-Monitor");
-        for (size_t i = 0; i < observedSignalCnt; i++) {
-            const observed_signal_t *signal = &observedSignals[i];
-            // SA_ONSTACK: handle the signal on separate stack
-            if (sigutils::install_handler(signal->signo, interceptor,
-                                          &signal->sa_previous, SA_ONSTACK)) {
-                _LOGI("Signal %d [%s] handler installed", signal->signo, signal->description);
-            } else {
-                _LOGE("Unable to install signal %d handler", signal->signo);
+        // The thread name is restricted to 16 chars, including the terminating null
+        if (0 != pthread_setname_np(pthread_self(), "NR-Sig-Handler")) {
+            _LOGE_POSIX("pthread_setname_np()");
+        } else {
+            for (size_t i = 0; i < observedSignalCnt; i++) {
+                const observed_signal_t *signal = &observedSignals[i];
+                if (sigutils::install_handler(signal->signo, interceptor,
+                                              &signal->sa_previous, SA_ONSTACK)) {
+                    _LOGI("Signal %d [%s] handler installed", signal->signo, signal->description);
+                } else {
+                    _LOGE("Unable to install signal %d handler", signal->signo);
+                }
             }
+
+            intercepting = 0;
+
+            _LOGI("Signal handler initialized");
         }
-
-        intercepting = 0;
-
-        _LOGI("Signal handler initialized");
     }
 
     return nullptr;
@@ -113,15 +125,13 @@ void *install_handler(__unused void *unused) {
  * Remove all installed signal interceptors
  */
 void uninstall_handler() {
-    if (initialized) {
+    if (initialized > 0) {
         for (size_t i = 0; i < observedSignalCnt; i++) {
             observed_signal_t signal = observedSignals[i];
             sigaction(signal.signo, &signal.sa_previous, nullptr);
-            memset(&signal.sa_previous, 0, sizeof(struct sigaction));
         }
 
         initialized--;
-        intercepting = 0;
 
         _LOGI("Signal handler uninstalled");
     }
@@ -131,8 +141,6 @@ void uninstall_handler() {
  * release all alloc'd heap:
  */
 void dealloc() {
-    _LOGD("dealloc called");
-
     if (_stack.ss_sp != nullptr) {
         free(_stack.ss_sp);
         _stack.ss_sp = nullptr;
@@ -141,30 +149,35 @@ void dealloc() {
 }
 
 bool signal_handler_initialize() {
-    pthread_mutex_lock(&mutex);
+    if (0 == pthread_mutex_lock(&mutex)) {
 
-    if (!sigutils::set_sigstack(&_stack, SIGSTKSZ * 2)) {
-        _LOGE("Signal handlers are disabled: could not set the handler signal stack");
-        return false;
-    }
-
-    // Main thread does not block SIGQUIT by default.
-    // Block it and start a new thread to handle all signals,
-    // using the signal mask of the parent thread.
-
-    if (sigutils::block_signal(SIGQUIT)) {
-        auto handlerThread = (pthread_t) nullptr;
-
-        // new thread will inherit the sigmask of the parent
-        if (pthread_create(&handlerThread, nullptr, install_handler, nullptr) != 0) {
-            _LOGE("Unable to create watchdog thread");
+        if (!sigutils::set_sigstack(&_stack, SIGSTKSZ * 2)) {
+            _LOGE("Signal handlers are disabled: could not set the handler signal stack");
+            return false;
         }
 
-        // restore SIGQUIT on this thread
-        sigutils::unblock_signal(SIGQUIT);
-    }
+        // Main thread does not block SIGQUIT by default.
+        // Block it and start a new thread to handle all signals,
+        // using the signal mask of the parent thread.
 
-    pthread_mutex_unlock(&mutex);
+        if (sigutils::block_signal(SIGQUIT)) {
+            auto handlerThread = (pthread_t) nullptr;
+
+            // new thread will inherit the sigmask of the parent
+            if (0 != pthread_create(&handlerThread, nullptr, install_signal_observers, nullptr)) {
+                _LOGE_POSIX("Unable to create monitor thread");
+            }
+
+            // restore SIGQUIT on this thread
+            sigutils::unblock_signal(SIGQUIT);
+        }
+
+        if (0 != pthread_mutex_unlock(&mutex)) {
+            _LOGE_POSIX("pthread_mutex_unlock()");
+        }
+    } else {
+        _LOGE_POSIX("pthread_mutex_lock() failed");
+    }
 
     return true;
 }
@@ -174,27 +187,32 @@ bool signal_handler_initialize() {
  */
 void signal_handler_shutdown() {
     _LOGI("Shutting down signal handler");
-    pthread_mutex_lock(&mutex);
-    uninstall_handler();
-    dealloc();
-    pthread_mutex_unlock(&mutex);
-    _LOGI("The signal handler has shutdown");
+    if (0 == pthread_mutex_lock(&mutex)) {
+        uninstall_handler();
+        dealloc();
+        if (0 == pthread_mutex_unlock(&mutex)) {
+            _LOGI("The signal handler has shutdown");
+            return;
+        }
+    }
+    _LOGE_POSIX("pthread_mutex_lock() failed");
 }
 
 /**
- * Invoke the proper handler for a passed signal
+ * Invoke the default handler for a passed signal
  */
-void invoke_sigaction(int signo, struct sigaction *_sigaction, siginfo_t *_siginfo, void *ucontext) {
+void
+invoke_sigaction(int signo, struct sigaction *_sigaction, siginfo_t *_siginfo, void *ucontext) {
     if (_sigaction->sa_flags & SA_SIGINFO) {
-        _LOGD("Signal %d: calling sigaction w/siginfo", signo);
+        _LOGD("Calling signal[%d] sigaction w/siginfo", signo);
         _sigaction->sa_sigaction(signo, _siginfo, ucontext);
 
     } else if (_sigaction->sa_handler == SIG_DFL) {
-        _LOGD("Signal %d: calling default handler", signo);
+        _LOGD("Calling signal[%d] default handler", signo);
         raise(signo);
 
     } else if (_sigaction->sa_handler != SIG_IGN) {
-        _LOGD("Signal %d: ignored", signo);
+        _LOGD("Signal [%d] action ignored", signo);
         void (*handler)(int) = _sigaction->sa_handler;
         handler(signo);
     }
@@ -204,15 +222,28 @@ void invoke_sigaction(int signo, struct sigaction *_sigaction, siginfo_t *_sigin
  * The process may be killed during this function execution and may never return
  */
 void invoke_previous_sigaction(int signo, siginfo_t *_siginfo, void *ucontext) {
-    pthread_mutex_lock(&mutex);
+    if (0 == pthread_mutex_lock(&mutex)) {
+        for (size_t i = 0; i < observedSignalCnt; ++i) {
+            observed_signal_t signal = observedSignals[i];
+            if (signal.signo == signo) {
+                _LOGI("Invoking previous handler for signal %d", signal.signo);
+                invoke_sigaction(signo, &signal.sa_previous, _siginfo, ucontext);
+            }
+        }
+        if (0 != pthread_mutex_unlock(&mutex)) {
+            _LOGE_POSIX("pthread_mutex_unlock() failed");
+        }
+    } else {
+        _LOGE_POSIX("pthread_mutex_lock() failed");
+    }
+}
 
+observed_signal_t *observed_signal_get_or_null(int signo) {
     for (size_t i = 0; i < observedSignalCnt; ++i) {
-        observed_signal_t signal = observedSignals[i];
-        if (signal.signo == signo) {
-            _LOGI("Invoking previous handler for signal %d", signal.signo);
-            invoke_sigaction(signo, &signal.sa_previous, _siginfo, ucontext);
+        if (observedSignals[i].signo == signo) {
+            return &observedSignals[i];
         }
     }
 
-    pthread_mutex_unlock(&mutex);
+    return nullptr;
 }
